requirements.txt
dask[dataframe]
ftfy
geopandas
geopy
google-cloud-bigquery[pandas]
matplotlib
pandas
pyarrow
pynecone
rapidfuzz
requests
seaborn
shapely
sqlalchemy-bigquery

# para modelos ML
huggingface_hub[cli]
ipywidgets
torch
ctransformers
#accelerate  # uso GPU/TPU

# para Reviews-bot
reflex
#openai

# uso en `tests/`
requests
yelp

#psycopg2-binary  # Uso en 'tests/'
#sec_api  # uso API de la SEC

